requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# HTTP requests for routing APIs
requests>=2.31.0

# Polyline encoding/decoding
polyline>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-django>=4.5.0

# Production WSGI server
gunicorn>=21.2.0

# Environment variable management
python-dotenv>=1.0.0
